continuous update of profile filled and form styles
